chore(wmg): refactor BottomBanner
