+++ frontend/src/components/BottomBanner/constants.ts
@@ -1,0 +1,8 @@
+export const FORM_CONTAINER_ID = "hubspot-form-container";
+export const FORM_CONTAINER_ID_QUERY = `#${FORM_CONTAINER_ID}`;
+export const FAILED_EMAIL_VALIDATION_STRING =
+  "Please provide a valid email address.";
+export const HIDDEN_NEWSLETTER_SUCCESS_MESSAGE =
+  "Thank you for joining our newsletter.";
+export const BOTTOM_BANNER_EXPIRATION_TIME_MS = 30 * 24 * 60 * 60 * 1000; // 30 days
+export const BOTTOM_BANNER_LAST_CLOSED_TIME_KEY = "bottomBannerLastClosedTime";
